% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_countries.R
\name{aqer_countries}
\alias{aqer_countries}
\title{Function to get Air Quality e-Reporting (AQER) countries.}
\usage{
aqer_countries()
}
\value{
Tibble.
}
\description{
Function to get Air Quality e-Reporting (AQER) countries.
}
\examples{

# Get countries
data_countries <- aqer_countries()

}
\author{
Stuart K. Grange
}
