% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_data.R
\name{aqer_data}
\alias{aqer_data}
\title{Function to import Air Quality e-Reporting (AQER) observational data.}
\usage{
aqer_data(url, encoding = "UCS-2LE", as_smonitor = FALSE,
  verbose = FALSE)
}
\arguments{
\item{url}{A vector of URLs from \code{\link{aqer_file_list}}.}

\item{encoding}{Encoding of \code{url}.}

\item{as_smonitor}{Should the return be formatted for use within the 
\code{smonitor} framework?}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to import Air Quality e-Reporting (AQER) observational data.
}
\examples{

# Example url
url <- "https://ereporting.blob.core.windows.net/downloadservice/GI_9_22755_2017_timeseries.csv"
data_aqer <- aqer_data(url)

}
\seealso{
\code{\link{aqer_file_list}}
}
\author{
Stuart K. Grange
}
