% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_data.R
\name{aqer_data_clean}
\alias{aqer_data_clean}
\title{Function to format Air Quality e-Reporting (AQER) observational data for easy
use in the \strong{smonitor} framework.}
\usage{
aqer_data_clean(df)
}
\arguments{
\item{df}{Data frame from \code{\link{aqer_data}}.}
}
\value{
Tibble.
}
\description{
Function to format Air Quality e-Reporting (AQER) observational data for easy
use in the \strong{smonitor} framework.
}
\examples{

# Example url
url <- "https://ereporting.blob.core.windows.net/downloadservice/GI_9_22755_2017_timeseries.csv"
data_aqer <- aqer_data(url)

# Clean data
data_aqer_clean <- aqer_data_clean(data_aqer)

}
\seealso{
\code{\link{aqer_data}}
}
\author{
Stuart K. Grange
}
