% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_file_list.R
\name{aqer_file_list}
\alias{aqer_file_list}
\title{Function to get remote file lists containing Air Quality e-Reporting (AQER)
observational data.}
\usage{
aqer_file_list(
  country = "",
  pollutant = "",
  start = 2013,
  end = NA,
  station = "",
  source = "all",
  verbose = FALSE
)
}
\arguments{
\item{country}{Country code to filter file list to.}

\item{pollutant}{Pollutant code to filter file list to.}

\item{start}{Start year to filter file list to.}

\item{end}{End year to filter file list to.}

\item{station}{Station code to filter file list to.}

\item{source}{What data source should be used? Can be either \code{"all"},
\code{"e1a"} for validated and reported data, or \code{"e2a"} for near real 
time data which are not validated data.}

\item{verbose}{Should the function give messages?}
}
\value{
Character vector.
}
\description{
Function to get remote file lists containing Air Quality e-Reporting (AQER)
observational data.
}
\examples{

# Get all swiss files
aqer_file_list(country = "ch")

# Get all validated/reported for Switzerland
aqer_file_list(country = "ch", source = "e1a")

# Get all uk files for 2018
aqer_file_list(country = "gb", start = 2018, end = 2018)


}
\author{
Stuart K. Grange
}
