% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_metadata.R
\name{aqer_metadata}
\alias{aqer_metadata}
\title{Function to import Air Quality e-Reporting (AQER) metadata.}
\usage{
aqer_metadata(file = NA, as_smonitor = FALSE)
}
\arguments{
\item{file}{File name or URL of metadata file. Do not use for "live" version.}

\item{as_smonitor}{Should the return be formatted for use within the 
\code{smonitor} framework?}
}
\value{
Tibble.
}
\description{
Function to import Air Quality e-Reporting (AQER) metadata.
}
\examples{

# Load aqer metadata
data_metadata <- aqer_metadata()

# Load aqer metadata and clean a bit
data_metadata_clean <- aqer_metadata(as_smonitor = TRUE)

}
\author{
Stuart K. Grange
}
