% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_pollutants.R
\name{aqer_pollutants}
\alias{aqer_pollutants}
\title{Function to get Air Quality e-Reporting (AQER) pollutants.}
\usage{
aqer_pollutants(file = NA)
}
\arguments{
\item{file}{File name or URL of metadata file. Do not use for "live" version.}
}
\value{
Tibble.
}
\description{
Function to get Air Quality e-Reporting (AQER) pollutants.
}
\examples{

# Get pollutants
data_pollutants <- aqer_pollutants()

}
\seealso{
\code{\link{aqer_metadata}}
}
\author{
Stuart K. Grange
}
