% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_read_csv.R
\name{aqer_read_csv}
\alias{aqer_read_csv}
\title{Function to read an Air Quality e-Reporting (AQER) observational data file.}
\usage{
aqer_read_csv(file, encoding = "UCS-2LE", as_smonitor = FALSE,
  verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names to read.}

\item{encoding}{Encoding of \code{file}.}

\item{as_smonitor}{Should the return be formatted for use within the 
\code{smonitor} framework?}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to read an Air Quality e-Reporting (AQER) observational data file.
}
\note{
To-do: Enhance to use \strong{readr}.
}
\author{
Stuart K. Grange
}
