% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_sites.R
\name{aqer_sites}
\alias{aqer_sites}
\title{Function to get Air Quality e-Reporting (AQER) sites/stations.}
\usage{
aqer_sites(as_smonitor = FALSE)
}
\arguments{
\item{as_smonitor}{Should the return be formatted for use within the 
\code{smonitor} framework?}
}
\value{
Tibble.
}
\description{
Function to get Air Quality e-Reporting (AQER) sites/stations.
}
\examples{

# Get countries
data_sites <- aqer_sites()

}
\author{
Stuart K. Grange
}
