% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e2a_data.R
\name{clean_e2a_data}
\alias{clean_e2a_data}
\title{Function to format Air Quality e-Reporting (AQER) observational data for easy
use in the \strong{smonitor} framework.}
\usage{
clean_e2a_data(df)
}
\arguments{
\item{df}{Data frame from \code{\link{get_e2a_data}}}
}
\value{
Tibble.
}
\description{
Function to format Air Quality e-Reporting (AQER) observational data for easy
use in the \strong{smonitor} framework.
}
\examples{

# Example url
url <- "https://ereporting.blob.core.windows.net/downloadservice/GI_9_22755_2017_timeseries.csv"
data_aqer <- get_e2a_data(url)

# Clean data
data_aqer_clean <- clean_e2a_data(data_aqer)

}
\seealso{
\code{\link{get_e2a_data}}
}
\author{
Stuart K. Grange
}
