% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e2a_countries.R
\name{get_e2a_countries}
\alias{get_e2a_countries}
\title{Function to get Air Quality e-Reporting (AQER) countries.}
\usage{
get_e2a_countries()
}
\value{
Tibble.
}
\description{
Function to get Air Quality e-Reporting (AQER) countries.
}
\examples{

# Get countries
data_countries <- get_e2a_countries()

}
\author{
Stuart K. Grange
}
