% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e2a_data.R
\name{get_e2a_data}
\alias{get_e2a_data}
\title{Function to import Air Quality e-Reporting (AQER) observational data.}
\usage{
get_e2a_data(url, as_smonitor = FALSE, verbose = FALSE)
}
\arguments{
\item{url}{A vector of URLs from \code{\link{get_e2a_file_list}}.}

\item{as_smonitor}{Should the return be formatted for use within the 
\code{smonitor} framework?}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to import Air Quality e-Reporting (AQER) observational data.
}
\examples{

# Example url
url <- "https://ereporting.blob.core.windows.net/downloadservice/GI_9_22755_2017_timeseries.csv"
data_aqer <- get_e2a_data(url)

}
\seealso{
\code{\link{get_e2a_file_list}}
}
\author{
Stuart K. Grange
}
