% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e2a_metadata.R
\name{get_e2a_metadata}
\alias{get_e2a_metadata}
\title{Function to import Air Quality e-Reporting (AQER) metadata.}
\usage{
get_e2a_metadata(file = NA, as_smonitor = FALSE)
}
\arguments{
\item{file}{File name or URL of metadata file. Do not use for "live" version.}

\item{as_smonitor}{Should the return be formatted for use within the 
\code{smonitor} framework?}
}
\value{
Tibble.
}
\description{
Function to import Air Quality e-Reporting (AQER) metadata.
}
\examples{

# Load metadata
data_metadata <- get_e2a_metadata()

}
\author{
Stuart K. Grange
}
