% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e2a_pollutants.R
\name{get_e2a_pollutants}
\alias{get_e2a_pollutants}
\title{Function to get Air Quality e-Reporting (AQER) pollutants.}
\usage{
get_e2a_pollutants()
}
\value{
Tibble.
}
\description{
Function to get Air Quality e-Reporting (AQER) pollutants.
}
\examples{

# Get countries
data_pollutants <- get_e2a_pollutants()

}
\author{
Stuart K. Grange
}
