% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e2a_sites.R
\name{get_e2a_sites}
\alias{get_e2a_sites}
\title{Function to get Air Quality e-Reporting (AQER) sites/stations.}
\usage{
get_e2a_sites(as_smonitor = FALSE)
}
\arguments{
\item{as_smonitor}{Should the return be formatted for use within the 
\code{smonitor} framework?}
}
\value{
Tibble.
}
\description{
Function to get Air Quality e-Reporting (AQER) sites/stations.
}
\examples{

# Get countries
data_sites <- get_e2a_sites()

}
\author{
Stuart K. Grange
}
