% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_data_clean.R
\name{aqer_data_clean}
\alias{aqer_data_clean}
\title{Function to format Air Quality e-Reporting (AQER) observational data for easy
use in the \strong{smonitor} framework.}
\usage{
aqer_data_clean(df, drop_sites = FALSE)
}
\arguments{
\item{df}{Tibble from \code{\link{aqer_read_csv}}.}

\item{drop_sites}{Should bad sites be removed from the table?}
}
\value{
Tibble.
}
\description{
Function to format Air Quality e-Reporting (AQER) observational data for easy
use in the \strong{smonitor} framework.
}
\examples{

# Example url
url <- "https://ereporting.blob.core.windows.net/downloadservice/GI_9_22755_2017_timeseries.csv"
data_aqer <- aqer_read_csv(url)

# Clean data
data_aqer_clean <- aqer_data_clean(data_aqer)

# Another example, this files has missing units and will raise a warning
url <- "https://ereporting.blob.core.windows.net/downloadservice/AT_618_1437_2013_timeseries.csv"
data_aqer_clean_test <- aqer_read_csv(url) \%>\% 
  aqer_data_clean()

}
\seealso{
\code{\link{aqer_read_csv}}
}
\author{
Stuart K. Grange
}
