% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_file_list.R
\name{aqer_file_list}
\alias{aqer_file_list}
\title{Function to get remote file lists containing Air Quality e-Reporting (AQER)
observational data.}
\usage{
aqer_file_list(country = "", pollutant = "", start = NA, end = NA,
  station = "", verbose = FALSE)
}
\arguments{
\item{country}{Country code to filter file list to.}

\item{pollutant}{Pollutant code to filter file list to.}

\item{start}{Start year to filter file list to.}

\item{end}{End year to filter file list to.}

\item{station}{Station code to filter file list to.}

\item{verbose}{Should the function give messages?}
}
\value{
Character vector.
}
\description{
Function to get remote file lists containing Air Quality e-Reporting (AQER)
observational data.
}
\examples{

# Get all swiss files
aqer_file_list(country = "ch")

# Get all uk files for 2018
aqer_file_list(country = "gb", start = 2018, end = 2018)


}
\author{
Stuart K. Grange
}
