% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_pollutants.R
\name{aqer_pollutants}
\alias{aqer_pollutants}
\title{Function to get Air Quality e-Reporting (AQER) pollutants.}
\usage{
aqer_pollutants()
}
\value{
Tibble.
}
\description{
Function to get Air Quality e-Reporting (AQER) pollutants.
}
\examples{

# Get pollutants
data_pollutants <- aqer_pollutants()

}
\author{
Stuart K. Grange
}
