% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqer_read_csv.R
\name{aqer_read_csv}
\alias{aqer_read_csv}
\alias{aqer_read_csv_worker}
\title{Functions to read Air Quality e-Reporting (AQER) observational data files.}
\usage{
aqer_read_csv(file, encoding = "UCS-2LE", as_smonitor = FALSE,
  verbose = FALSE)

aqer_read_csv_worker(file, encoding, as_smonitor, verbose)
}
\arguments{
\item{file}{Vector of file names to read.}

\item{encoding}{Encoding of \code{file}.}

\item{as_smonitor}{Should the return be formatted for use within the 
\code{smonitor} framework?}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
\code{aqer_read_csv_worker} is the unvectorised worker function which can be
used other situations other than just reading many files, such as exporting 
data in parallel.
}
\note{
To-do: Enhance to use \strong{readr}, but need to address encoding 
issues first.
}
\examples{

# Example url
url <- "https://ereporting.blob.core.windows.net/downloadservice/GI_9_22755_2017_timeseries.csv"
data_aqer <- aqer_read_csv(url)

# Correct some data types
data_aqer_clean <- aqer_read_csv(url, as_smonitor = TRUE)

}
\seealso{
\code{\link{aqer_file_list}}
}
\author{
Stuart K. Grange
}
